fixed Total Return metric
